% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imarpe-data.R
\docType{data}
\name{AIPShapefile}
\alias{AIPShapefile}
\title{Isoparalitoral area.}
\format{A \code{shapefile} with the information of the isoparalitoral area: longitude, latitude, area and perimeter.}
\usage{
AIPShapefile
}
\description{
The peruvian shape isoparalitoral area.
}
\references{
imarpe package vignette.
}
