% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{assignAnchovy_season}
\alias{assignAnchovy_season}
\title{Title Assing anchovy season}
\usage{
assignAnchovy_season(x, region)
}
\arguments{
\item{x}{A vector of date (with day, month and year) in the format \code{Date}.}

\item{region}{A object indicating if is the north-central region ('norte-centro') or if is
the south region ('sur').}
}
\value{
A object with the season code.
}
\description{
Using the historical records of the opening and closing of
anchovy fishing by regions (north-central and south), this function can identify the
anchovy season.
}
\examples{
# Construct a vector of dates

exampleDates = as.Date(c("20/01/2013", "30/01/1999"), format = "\%d/\%m/\%Y")

assignAnchovy_season(exampleDates, region = "norte-centro")

}
