% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imarpe-auxiliar.R
\name{capitalizeFirstLetter}
\alias{capitalizeFirstLetter}
\title{Capitalize first letter}
\usage{
capitalizeFirstLetter(x)
}
\arguments{
\item{x}{A R object with elements to the character class.}
}
\value{
A object with the first letter of each element on capital letter.
}
\description{
Function that capitalize only the first letter of a object and convert the
rest of the object on lower case.
}
\examples{
capitalizeFirstLetter(x = c("the LAsT BOOK", "my favoURIte MOVIe"))

}
