% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{catchSpecies}
\alias{catchSpecies}
\title{Species catch in the database of PBP.}
\format{A \code{data.frame} with three columns of: name of spcies catch, taxonomic group, and the common species name.}
\usage{
catchSpecies
}
\description{
Data with a list of species and taxonomic groups finding on bitacoras catches.
}
\references{
imarpe package vignette.
}

