% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery-class.R
\name{combineFisheryVar}
\alias{combineFisheryVar}
\title{Combine variables for fishery class}
\usage{
combineFisheryVar(landing, effort)
}
\arguments{
\item{landing}{complete this.}

\item{effort}{complete this}
}
\value{
x object.
}
\description{
complete this.
}

