% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitacora-internal.R
\name{effortSpeciesData.bitacora}
\alias{effortSpeciesData.bitacora}
\title{Get species effort data}
\usage{
effortSpeciesData.bitacora(data, species, region = NULL)
}
\arguments{
\item{data}{A \code{data.frame} obtained from effortData.bitacora function (internal function
of bitacora class). This data frame could be obtained from
getMainResults.bitacora function too.}

\item{species}{A \code{character} with the common name of anchovy ("anchoveta"),
sardine ("sardina"), jack mackerel ("jurel"), chub mackerel ("caballa") and bonito ("bonito").}

\item{region}{If the species is anchovy (\code{anchoveta}) is necessary
indicate for each region the effort data is going to be calculate. By default is \code{NULL}.}
}
\value{
A \code{data.frame} with effort data for a given species.
}
\description{
Receive an object of bitacora class, indexes the data by species,
estimates the season by species, for anchovy the season depend of region
(north-central and south region) but for other species
(sardine, jack mackerel, chub mackerel and bonito) the season is annual.
}
\examples{
# Read a example of data base
bitacoraData = system.file("extdata", "bitacoraData.csv", package = "imarpe")

# Create a object of bitacora class
bitacoraObject = getBitacoraData(file = bitacoraData)
class(bitacoraObject)

# Get the results
mainBitacoraData = getMainResults.bitacora(object = bitacoraObject, language = "spanish",
specie = "anchoveta", effortData = TRUE)

# Effort data
effortData = mainBitacoraData$effortData

# Use of effortSpeciesData.bitacora
effortNC = effortSpeciesData.bitacora(data = effortData, species = "anchoveta", region = "norte-centro")

}
\author{
Criscely Lujan-Paredes, \email{criscelylujan@gmail.com}.
}
