% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{engToSpa}
\alias{engToSpa}
\title{Months in english to months in spanish}
\usage{
engToSpa(x)
}
\arguments{
\item{x}{A vector with abbreviation of months in english.}
}
\value{
A vector with abbreviation of months in spanish.
}
\description{
Function to convert abbreviation of months in english to abbreviation of
months in spanish.
}
\examples{
# The 'months' vector contain abbreviation of months in english.
months = month.abb
engToSpa(x = months)

}

