% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fisheryData}
\alias{fisheryData}
\title{Fishery data}
\format{A \code{data.frame}.}
\usage{
data(fisheryData).
}
\description{
A data base (\code{data.frame}) with fishery data about a
species. This data base contains information about the landing and the effort
 of the species. The variables appear on the columns and are:
 \itemize{
 \item Anho for the years.
 \item Mes for the months.
 \item Dia for the days.
 \item Especies for a species.
 \item Tipo_flota for the types of fleets.
 \item Puerto for the ports.
 \item Longitud for longitude.
 \item Latitud for latitude.
 \item Captura for the landings.
 \item Capacidad_Bodega for storage capacity.
 \item Anzuelos for the hook number.
 \item Embarcaciones for number of boats.
 }
}

