% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imarpe-auxiliar.R
\name{getAIPInfo}
\alias{getAIPInfo}
\title{Title Extract information from isoparalitoral area code}
\usage{
getAIPInfo(aipVector)
}
\arguments{
\item{aipVector}{A vector indicating the AIP code.}
}
\value{
A \code{data.frame} with variables distance to the coast (dc), latitude (lat)
and position (upDown) in the columns for each AIP code in rows.
}
\description{
The isoparalitoral area (AIP) code must be written in format DDLLPP,
where DD is value of distance to coast (mn/10), LL are values of latitude (as integer) and
PP is position (up or down).
}
\examples{
getAIPInfo(c(30073, 1020, 2010))

}
