% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{getCalendar}
\alias{getCalendar}
\title{Get calendar}
\usage{
getCalendar(year, language = "spanish")
}
\arguments{
\item{year}{The specific year to get the data base of dates.}

\item{language}{The language for columns names in the data frame that is returned from the function. By default
is in spanish, but this can be english.}
}
\value{
A data frame with three columns, the years, months and days.
}
\description{
Function to get a data base of dates according a specific year.
}
\examples{
#To get the a data base of dates to the year 2001 in spanish.
getCalendar(year = 2001)

#To get the same data base of dates but with the colnames in englis, change the \\code{language} parameter to 'english'.
getCalendar(year = 2001, language = "english")

}

