% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitacora-internal.R
\name{getCpue_relative}
\alias{getCpue_relative}
\title{Get the relative catch per unit effort (cpue)}
\usage{
getCpue_relative(data, toTons = FALSE, efforType, cpueBy = "time", timeBy,
  fleeType = NULL)
}
\arguments{
\item{data}{A \code{data.frame} with information about catchs and fishing effort.}

\item{toTons}{A logical parameter. \code{TRUE} (dafault) it assume that the catches is in kilograms
and converts it into tonnes (divided by 1000). \code{FALSE} it assume the the information is in tons
and don't convert the data.}

\item{efforType}{A \code{character} indicating the fishing effort type. It could be four types:
\itemize{
  \item travelTime for the effort as a function of the travel time.
  \item haulTotal for the effort as a function of the total number of fishing haul.
  \item storageCapacity for the effort as a function of the storage capacity.
  \item searchTime for the effort as a function of the search time to catch the species.
}}

\item{cpueBy}{Parameter to indicate whether the relative cpue will be estimated by
time (\code{cpueBy = time}) or by port (\code{cpueBy = port}).}

\item{timeBy}{If the relative cpue is estimated over the time, it could be by days
(\code{timeBy = days}), months (\code{timeBy = months}), years (\code{timeBy = years}) or
by seasons (\code{timeBy = seasons}).}

\item{fleeType}{The relative cpue could be estimated for the whole data given or for a data by fleeType.
By default this parameter is \code{NULL}.}
}
\value{
A \code{vector} with the relative cpue.
}
\description{
Get the relative catch per unit effort (cpue)
}
\details{
The relative cpue is a index used on bitacora report.
}

