% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getDailyReport}
\alias{getDailyReport}
\title{Title}
\usage{
getDailyReport(directory = NULL,
  urlFishingMonitoring = "http://www.imarpe.pe/imarpe/archivos/reportes/imarpe_rpelag_porfinal",
  datesList, simpleFreqSizes, dataCruise, officialBiomass, threshold = 30,
  species = "Anchoveta", a, b, growthParameters = list(k = 0.83, Linf =
  19.21, sizeM = c(0, 8, 12), vectorM = rep(0.8, 3), catchFactor = 1))
}
\arguments{
\item{officialBiomass}{}
}
\description{
Title
}

