% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitacora-internal.R
\name{getEffort}
\alias{getEffort}
\title{Get fishing effort}
\usage{
getEffort(data, efforType, effortBy, timeBy, fleeType = NULL)
}
\arguments{
\item{data}{A \code{data.frame} with information to estimated the fishing effort (in columns).
This data base have to contain columns of the time (days, months, years, season), ports and the
fleet type.}

\item{efforType}{A \code{character}. It could be one of the fishing effort type ("travelTime",
"haulTotal", "storageCapacity", "searchTime").}

\item{effortBy}{Parameter to indicate whether the effort will be estimated by
time (\code{effortBy = time}) or by port (\code{effortBy = port}).}

\item{timeBy}{If the effort is estimated over the time, it could be by days
(\code{timeBy = days}), months (\code{timeBy = months}), years (\code{timeBy = years}) or
by seasons (\code{timeBy = seasons}).}

\item{fleeType}{The effort could be estimated for the whole data given or for a data by fleeType.
By default this parameter is \code{NULL}.}
}
\value{
A \code{vector} with the estimated effort.
}
\description{
Generical function to get the fishing effort. The type of fishing effort could be four types:
\itemize{
  \item travelTime for the effort as a function of the travel time.
  \item haulTotal for the effort as a function of the total number of fishing haul.
  \item storageCapacity for the effort as a function of the storage capacity.
  \item searchTime for the effort as a function of the search time to catch the species.
}
}
\examples{
# Read a example of data base
bitacoraData = system.file("extdata", "bitacoraData.csv", package = "imarpe")

# Create a object of bitacora class
bitacoraObject = getBitacoraData(file = bitacoraData)

# Get the results
mainBitacoraData = getMainResults.bitacora(object = bitacoraObject, language = "spanish",
specie = "anchoveta", effortData = TRUE)

# Effort data
effortData = mainBitacoraData$effortData

effortNC = effortSpeciesData.bitacora(data = effortData, species = "anchoveta", region = "norte-centro")

# Get the cpue
effortNC_tt = getEffort(data = effortNC, efforType = "travelTime", effortBy = "port")
effortNC_hn = getEffort(data = effortNC, efforType = "haulTotal", effortBy = "port")

}
\author{
Criscely Lujan-Paredes, \email{criscelylujan@gmail.com}.
}
