% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imarpe-main.R
\name{getMainResults.bitacora}
\alias{getMainResults.bitacora}
\title{Get main results from bitacora}
\usage{
getMainResults.bitacora(object, observedTrip = NULL, fishingHaul = NULL,
  fishingPoints = NULL, speciesComposition = NULL,
  distributionCatch = NULL, effortData = NULL, language = "spanish",
  latByPort = FALSE, specie = "anchoveta")
}
\arguments{
\item{object}{Object of \code{bitacora} class.}

\item{observedTrip}{Parameter to indicate whether the observed trip will be estimated.
By default is \code{NULL} but if it will be estimated receive the logical value of \code{TRUE}.}

\item{fishingHaul}{Parameter to indicate whether the fishing haul sampled will be
estimated. By default is \code{NULL} but if it will be estimated receive the
logical value of \code{TRUE}.}

\item{fishingPoints}{Parameter to indicate whether the fishing poins will be estimated.
By default is \code{NULL} but if it will be estimated received the logical value of \code{TRUE}.}

\item{speciesComposition}{Parameter to indicate whether the species composition will be
estimated. By default is \code{NULL} but if it will estimated received the logical value of \code{TRUE}.}

\item{distributionCatch}{Parameter to indicate whether the distribution of catches will be
estimated. By default is \code{NULL} but if it will estimated received the logical value of \code{TRUE}.}

\item{effortData}{Parameter to indicate whether the effort data will be
estimated. By default is \code{NULL} but if it will estimated received the logical value of \code{TRUE}.}

\item{language}{The select language to print the results. By default is "spanish".}

\item{latByPort}{\code{logical}. Parameter of fishingHaul.bitacora function. By
default is \code{FALSE} indicating that latitude information is not get from
port information. When is \code{TRUE}, the latitude is get from port information.}

\item{specie}{\code{character}. Parameter of distributionCatch.bitacora function.
Receives the name of the species that is calculated the distribution of the catch.
Default value is "anchoveta" but this could be:
\itemize{
  \item "anchoveta" to estimated the anchovy catches distribution.
  \item "sardina" to estimated the sardine catches distribution.
  \item "jurel" to estimated the jack mackerel catches distribution.
  \item "caballa" to estimated the chub mackerel catches distribution.
  \item "bonito" to estimated the bonito catches distribution.
}}
}
\value{
A list of \code{bitacora_mainResults} class. The \code{length} of the list is six,
one by each parameter (observedTrip, fishingHaul, fishingPoints, speciesComposition,
distributionCatch, and effortData).
}
\description{
Principal function of bitacora class objects to get principal results.
}
\details{
If one of the parameter (observedTrip, fishingHaul, fishingPoints, speciesComposition,
distributionCatch, and effortData) is \code{NULL} on \code{getMainResults.bitacora} function
the output of this parameter on the list produced by the function is \code{NULL} too.
}
\examples{
# Read a example of data base
bitacoraData = system.file("extdata", "bitacoraData.csv", package = "imarpe")

# Create a object of bitacora class
bitacoraObject = getBitacoraData(file = bitacoraData)
class(bitacoraObject)

# Get the results
mainBitacoraData = getMainResults.bitacora(object = bitacoraObject, language = "spanish",
specie = "anchoveta", observedTrip = TRUE, fishingHaul = TRUE, distributionCatch = TRUE)

# See the principal results bitacora class object
mainBitacoraData$observedTrip
mainBitacoraData$fishingHaul
mainBitacoraData$distributionCatch

}
\author{
Criscely Lujan-Paredes, \email{criscelylujan@gmail.com}.
}
