% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{getPort}
\alias{getPort}
\title{Get port information from name}
\usage{
getPort(myPorts)
}
\arguments{
\item{myPorts}{A vector of the names of ports.}
}
\value{
A list, with a list of the information about the ports and a vector with the
positions of the ports given to the function.
}
\description{
This function uses a data base with the list of ports around Peruvian
coast (portData) and according to names of the ports that is given, the function get information
about: the name of the port, the pattern to find the name of the port, the
longitude and latitude of their names, the area (north, south and central) of
ports and the importance of the port according the history of the fishery.
}
\examples{
# The function receive a vector with port names:
ports = getPort(c("T. de mora", "Paita"))
ports$data
ports$position

}

