% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{imarpe}
\alias{imarpe}
\alias{imarpe-package}
\title{imarpe: R package for the the automation of graphs, tables and reports of the Instituto del Mar del Peru}
\description{
This packages provides tools for the elaboration of graphs and tables that are
routinely performed in reports produced by Instituto del Mar del Peru. The package imarpe
is designed to work with generic functions regardless of the type of data used.
}
\author{
Criscely Lujan-Paredes \email{criscelylujan@gmail.com}, Luis Lau-Medrano \email{luis.laum@gmail.com}.
}
\keyword{automatization,}
\keyword{fishery,}
\keyword{imarpe.}
\keyword{reports,}

