% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imarpe-data.R
\docType{data}
\name{imarpeColours}
\alias{imarpeColours}
\title{Colours for the package.}
\usage{
imarpeColours
}
\description{
Vector of colours to use in the package.
}
\references{
imarpe package vignette.
}
