\name{imarpe-package}
\alias{imarpe-package}
\alias{imarpe}
\docType{package}
\title{
imarpe: automatización de gráficas, tablas y reportes del Instituto del Mar del Perú.
}
\description{
Este paquete proporciona herramientas para la elaboración de gráficas y tablas que se realizan de manera rutinaria en los informes producidos por el Instituto del Mar del Perú (IMARPE). El paquete está diseñado para trabajar con funciones genéricas (e.g. \code{plot}, \code{summary}, \code{print}) sin importar el tipo de datos utilizado, facilitando su uso y permitiendo su utilización incluso con conocimientos mínimos de \code{R}. 
}
\details{
\tabular{ll}{
Package: \tab imarpe\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-03-11\cr
License: \tab TBD\cr
}

}
\author{
Criscely Luján Paredes, Ricardo Oliveros-Ramos

Maintainer: Ricardo Oliveros-Ramos <roliveros@imarpe.gob.pe>
}
\references{
Luján-Paredes C, Oliveros-Ramos R. 2014. Introducción al paquete \code{imarpe} para la automatización de gráficos, tablas y reportes usando \code{R}.
}
\keyword{ figuras, tablas, reportes, imarpe, R. }
\seealso{
}
\examples{
~~ simple examples of the most important functions ~~
}
