% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{isopArea.assigner}
\alias{isopArea.assigner}
\title{Title Get the isoparalitoral area code}
\usage{
isopArea.assigner(dataPoints, colLon = "lon", colLat = "lat")
}
\arguments{
\item{dataPoints}{\code{data.frame} which has longitude and latitude information.}

\item{colLon}{Name of column which contains longitude information.}

\item{colLat}{Name of column which contains latitude information.}
}
\value{
A numeric vector indicating the AIP values for each geographical coordinates.
}
\description{
Using information of the geographical coordinates calculate the
isoparalitoral area code (AIP).
}
\examples{
# Create a data frame with information of geographical coordinates.
exampleCoords = data.frame(lon = runif(n = 10, min = -80, max = -78),
                            lat = runif(n = 10, min = -14, max = -12))

isopArea.assigner(dataPoints = exampleCoords)

}
