% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{month.abb_spanish}
\alias{month.abb_spanish}
\title{Abbreviations of the months.}
\format{A \code{vector} containing the abbreviations of the months of the year.}
\usage{
month.abb_spanish
}
\description{
Vector with the months names abbreviations on spanish.
}
\references{
imarpe package vignette.
}
