% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{month2word}
\alias{month2word}
\title{Months in numbers to months in words}
\usage{
month2word(x)
}
\arguments{
\item{x}{A vector with months in numbers.}
}
\value{
A vector with the abbreviations of months in words. Where the number one
is "Jan" (January) and increasing until the number twelve that is "Dec" (December).
}
\description{
Function to convert months in number to abbreviations of months in words.
}
\examples{
# The 'months' vector have the numbers between one and twelve.
months = seq(from = 1, to = 12, by = 1)
month2word(x = months)

}
