% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.bitacoras}
\alias{plot.bitacoras}
\title{Plot a Bitacoras Object}
\usage{
\method{plot}{bitacoras}(x, type, detailed = FALSE, ...)
}
\arguments{
\item{x}{Object of class \dQuote{bitacoras}.}

\item{type}{Type of graph to plot. Possible types are:
\itemize{
  \item \dQuote{effort} for plotting a barplot of average travel time with a line of average coves by port.
  \item \dQuote{depth} for plotting a boxplot of depths by ranges of latitudes.
}}

\item{detailed}{Logical attribute for \dQuote{depth} \code{type}.
If it is TRUE, points will be added in the boxplot (FALSE by defautl).}

\item{\dots}{Other graphical parameters accepted by \code{\link{plot}}.}
}
\value{
None (invisible NULL).
}
\description{
The \code{plot} method for objects of class \dQuote{bitacoras}.
This function plots two kind of graphs, which are defined by \dQuote{type} argument.
}
\details{
In \dQuote{effort} graph, plotted statistics like travel time, number of coves are
average data of the gathered profiles by port.
}
\examples{
## Loading the data
data(Bitacoras)

## Plotting by effort
plot(Bitacoras,type="effort")

## Plotting by depth
plot(Bitacoras,type="depth")
}
\author{
Erick Chacon Montalvan, Vilma Romero Romero, Criscely Lujan Paredes, Ricardo Oliveros-Ramos.

Maintainer: Ricardo Oliveros-Ramos <\email{roliveros@imarpe.gob.pe}>
}
\references{
Lujan Paredes C., Oliveros-Ramos R., Chacon Montalvan E., Romero Romero V., 2014.
Introduction to \pkg{imarpe} package for the automation of graphs, charts and reports using \code{R}.
}
\seealso{
\code{\link{getData}}, \code{\link{print.bitacoras}}, \code{\link{summary.bitacoras}}
}

