% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-class.R
\name{plot.cpue}
\alias{plot.cpue}
\title{Plot method for cpue objects}
\usage{
\method{plot}{cpue}(x, language, ploType = NULL, daysToPlot = c(1, 8, 15,
  22), textAxis2 = NULL, textAxis4 = NULL, ...)
}
\arguments{
\item{x}{Object of \code{cpue} class.}

\item{language}{\code{character}. Define the language of text labels in plots.
It could be \code{"spanish"} or \code{"english"}.}

\item{ploType}{What type of plot should be draw. Possible types are:
\itemize{
  \item plotDaily for daily plot
  \item plotMonthly for monthly plot
  \item plotYearly for yearly plot
  \item plotNC to plot the north-central region
  \item plotS to plot the south region
}}

\item{daysToPlot}{if is a daily plot by default the x axis plot the first day of the month
(1, 8, 15, 22). This is including in a vector form.}

\item{textAxis2}{The text of the x axis.}

\item{textAxis4}{The text of the y axis.}

\item{...}{Extra arguments.}
}
\value{
A graph of the specified type in \code{ploType}.
}
\description{
This method takes a \code{cpue} object and make useful plots.
The plots can be daily, monthly, yearly or for north-central and south peruvian region.
}

