% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery-class.R
\name{plot.fishery}
\alias{plot.fishery}
\title{Plot method for fishery objects}
\usage{
\method{plot}{fishery}(x, language = "spanish", ploType = NULL,
  daysToPlot = c(1, 8, 15, 22), textAxis2 = NULL, textAxis4 = NULL,
  colBar = "gray", colLine = "red", ...)
}
\arguments{
\item{x}{Object of \code{fishing} class.}

\item{language}{\code{character}. Define the language of text labels in plots.
It could be \code{"spanish"} or \code{"english"}.}

\item{ploType}{What type of plot should be draw. Possible types are:
\itemize{
  \item plotDaily for daily plot
  \item plotMonthly for monthly plot
  \item plotYearly for yearly plot
  \item plotNC to plot the north-central region
  \item plotS to plot the south region
}}

\item{daysToPlot}{If is a daily plot by default the x-axis show the first day of a week
(1, 8, 15, 22) but could be change to show a specific day or to show the all days of the
data with \code{all}. This is including in a vector form.}

\item{textAxis2}{The text of the x axis.}

\item{textAxis4}{The text of the y axis.}

\item{...}{Extra arguments.}
}
\value{
A graph of the specified type in \code{ploType}.
}
\description{
This method takes a \code{fishery} object and make useful plots for
each variables (lading and fishing effort). The plots can be daily, monthly, yearly or
for north-central and south peruvian region.
}
\examples{
# Read a example of a data base
fisheryData = system.file("extdata", "fisheryData.csv", package = "imarpe")

# Produce a object of fishery class
landing  = getFishingData(file = fisheryData, type = "fisheryinfo", varType = "landing", sp = "caballa")
class(landing)

plot(landing)
plot(landing, daysToPlot = "2", colBar = "red")
plot(landing, language= "english")
plot(landing, ploType = "plotMonthly")
plot(landing, ploType = "plotMonthly", language= "english")
plot(landing, ploType = "plotYearly")
plot(landing, ploType = "plotYearly", language= "english", colBar = "green")
plot(landing, ploType = "plotPERU")
plot(landing, ploType = "plotNC")
plot(landing, ploType = "plotS")

}

