% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.landings}
\alias{plot.landings}
\title{Plot a Landings Object}
\usage{
\method{plot}{landings}(x, time = NULL, ...)
}
\arguments{
\item{x}{Object of class "landings".}

\item{time}{Required period of time. Possible types are
\itemize{
  \item "day" for plotting by days,
  \item "month" for plotting by months,
  \item "year" for plotting by years.
}}

\item{\dots}{Other graphical parameters accepted by \code{\link{plot}}.}
}
\value{
None (invisible NULL).
}
\description{
The \code{plot} method for objects of class \dQuote{landings}.
This function plots three kinds of graphs, which are defined by \dQuote{time} argument.
}
\details{
In \dQuote{day} graph, the user can define the extent of the days that he wants to plot by adding the start and end date parameters as a character class.
}
\examples{
## Loading the data
data(Landings)

## Plotting by day
# to plot the whole range of days
plot(Landings,time="day")
# to plot a subset of the days
plot(Landings,time="day",start="2013-02-14",end="2013-06-16")

## Plotting by month
plot(Landings,time="month")

## Plotting by year
plot(Landings,time="year")
}
\author{
Erick Chacon Montalvan, Vilma Romero Romero, Criscely Lujan Paredes, Ricardo Oliveros-Ramos.

Maintainer: Ricardo Oliveros-Ramos <\email{roliveros@imarpe.gob.pe}>
}
\references{
Lujan Paredes C., Oliveros-Ramos R., Chacon Montalvan E., Romero Romero V., 2014.
Introduction to \pkg{imarpe} package for the automation of graphs, charts and reports using \code{R}.
}
\seealso{
\code{\link{getData}}, \code{\link{print.landings}}, \code{\link{summary.landings}}
}

