% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotEffort}
\alias{plotEffort}
\title{PlotEffort method}
\usage{
plotEffort(effort1, effort2, ...)
}
\arguments{
\item{...}{Extra arguments passed to \code{plotEffort.numeric} function.}

\item{x}{Object of \code{bitacora} class.}
}
\description{
Method for plotEffort.numeric function.
}
\details{
For more details read the help of \code{\link{plotEffort.numeric}}.
}
\author{
Criscely Lujan-Paredes, \email{criscelylujan@gmail.com}.
}
