% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitacora-class.R
\name{plotFishingPoints.bitacora}
\alias{plotFishingPoints.bitacora}
\title{Plot for fishing points}
\usage{
\method{plotFishingPoints}{bitacora}(x, language = "spanish", dataType,
  colMap = "khaki1", colFleet = c("red", "blue", "green", "black"),
  cexPointCatch = FALSE, cexPoint = 0.8, cex.axis = 1.2, cexPorts = 0.9,
  ...)
}
\arguments{
\item{x}{Object of \code{bitacora} class.}

\item{language}{\code{character}. Define the language of text labels in plots, by default is \code{"spanish"}.}

\item{dataType}{To indicate the fishing points of the species to be graphed. This could be:
\itemize{
  \item "dataAnch" to graphic fishing points of anchovy
  \item "dataSar" to graphic fishing points of sardine
  \item "dataJur" to graphic fishing points of jack mackerel
  \item "dataCab" to graphic fishing points of chub mackerel
  \item "dataBon" to graphic fishing points of bonito
  \item "dataGroups" to graphic fishing points for other species differences to the five mentioned
  \item "dataTotal" to graphic fishing points of all species
}}

\item{colMap}{Assigns the color to the land domain. By default is \code{khaki1}.}

\item{colFleet}{A vector. Colouts assigned to each fleet type.
\itemize{
  \item "red" for artisanal type fleet (boats with storage capacity between 0 and 10 tons).
  \item "blue" for smaller scale type fleet (boats with storage capacity between 10 and 32.5 tons).
  \item "green" for industrial wood fleet (boats with storage capacity between 32.5 and 110 tons).
  \item "black" for industrial fleet (boats with storage capacity greather than 110 tons).
}}

\item{cexPointCatch}{\code{logical}. To plot the size of the fishing points according the catch size
\code{TRUE} or to plot the fishing points without considering the size of the catch \code{FALSE}
(by default).}

\item{cexPoint}{The size of the fishing points on the map. By default is 0.8.}

\item{cex.axis}{The size of the axis on the map. By default is 1.2.}

\item{cexPorts}{The size of the port names on the map. By default is 0.9.}

\item{...}{Extra arguments.}
}
\value{
A map for fishing points of the data type selected on \code{dataType}.
}
\description{
This method takes a \code{bitacora} object and plots fishing points.
}
\examples{
# Read a example of data base
bitacoraData = system.file("extdata", "bitacoraData.csv", package = "imarpe")

# Create a object of bitacora class
bitacoraObject = getBitacoraData(file = bitacoraData)
class(bitacoraObject)

# Use the parameter dataType = "dataAnch" to graph anchovy fishing points
plotFishingPoints(x = bitacoraObject, language = "spanish", dataType = "dataAnch")

# Use the parameter dataType = "dataCab" to graph chub mackerel fishing points
plotFishingPoints(x = bitacoraObject, language = "spanish", dataType = "dataCab")

}
