% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotFishingPoints}
\alias{plotFishingPoints}
\title{PlotFishignPoints method}
\usage{
plotFishingPoints(x, laguage, dataType, ...)
}
\arguments{
\item{x}{Object of \code{bitacora} class.}

\item{...}{Extra arguments passed to \code{plotFishingPoints.bitacora} function.}
}
\description{
Method for plotFishingPoins.bitacora function.
}
\details{
For more details read the help of \code{\link{plotFishingPoints.bitacora}}.
}
\author{
Criscely Lujan-Paredes, \email{criscelylujan@gmail.com}.
}
