% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitacora-class.R
\name{plotFishingPresence.bitacora}
\alias{plotFishingPresence.bitacora}
\title{Plot for fishing presence points}
\usage{
\method{plotFishingPresence}{bitacora}(x, byGroup = TRUE, group = NULL,
  colMap = "khaki1", colSpecies = NULL, colLegend = NULL, cexPoint = 1,
  cex.axis = 1.2, cexPorts = 0.9, cexLegend = 1, ...)
}
\arguments{
\item{x}{Object of \code{bitacora} class.}

\item{byGroup}{\code{logical}. To indicate if the fishing presence points will be plotted by species that
make up a taxonomic group (\code{TRUE}), or to plot the taxonomic group in general (\code{FALSE}) except
four: "neritico", "transzonal", "oceanico", "demersal".}

\item{group}{Indicates what taxonomic group should be graph on the map of fishing presence.
By defaul is \code{NULL} when \code{byGroup = FALSE}, but when \code{byGroup = TRUE} this parameter receives
the name of the taxonomic group. These might be: "neritico", "transzonal", "costero", "oceanico", "demersal",
"bentonico", "mesopelagico", "depredador", "elasmobranquio", "cefalopodo", "medusa", "crustaceo", "tunicado",
 "alga", "copepodo", "eufausido", "molusco", "equinodermo", "invertebrado", "dinogelado".}

\item{colMap}{Assigns the color to the land domain. By default is \code{khaki1}.}

\item{colSpecies}{A vector with colours that represent the non-tarjet species. By defaul is \code{NULL} and
use a internal \code{\link{vectorColours}}.}

\item{colLegend}{By default (\code{colLegend = NULL}) receives the vector with colours of \code{colSpecies}.
But this parameter could be changed including a vector of colour by each species.}

\item{cexPoint}{The size of the fishing points on the map. By default is 1.}

\item{cex.axis}{The size of the axis on the map. By default is 1.2.}

\item{cexPorts}{The size of the port names on the map. By default is 0.9.}

\item{cexLegend}{The size of the common name of species.}

\item{...}{Extra arguments.}
}
\value{
A map for fishing presence points by taxonomic group or in general (for all species if the data).
}
\description{
This method takes a \code{bitacora} object and plot on a map the presence of the
non-tarjet species for the fishery.
}
\examples{
# Read a example of data base
bitacoraData = system.file("extdata", "bitacoraData.csv", package = "imarpe")

# Create a object of bitacora class
bitacoraObject = getBitacoraData(file = bitacoraData)
class(bitacoraObject)

# Use the parameter byGroup = TRUE and then use the parameter group to indicate which taxonomic
# groups is going to be plotted:
plotFishingPresence(x = bitacoraObject, byGroup = TRUE, group = "neritico")
plotFishingPresence(x = bitacoraObject, byGroup = TRUE, group = "transzonal")

# Use the parameter byGroup = FALSE:
plotFishingPresence(x = bitacoraObject, byGroup = FALSE)

}

