% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotFishingPresence}
\alias{plotFishingPresence}
\title{PlotFishingPresence method}
\usage{
plotFishingPresence(x, ...)
}
\arguments{
\item{x}{Object of \code{bitacora} class.}

\item{...}{Extra arguments passed to \code{plotFishingPresence.bitacora} function.}
}
\description{
Method for plotFishingPresence.bitacora function.
}
\details{
For more details read the help of \code{\link{plotFishingPresence.bitacora}}.
}

