% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitacora-class.R
\name{plotSpeciesComposition.bitacora}
\alias{plotSpeciesComposition.bitacora}
\title{Plot for species composition of catch}
\usage{
\method{plotSpeciesComposition}{bitacora}(x, threshold = TRUE,
  minPercentage = 0.2, ...)
}
\arguments{
\item{x}{Object of \code{bitacora} class.}

\item{threshold}{Logical, by default is \code{TRUE} when the species composition on the
catches will be graphic according to a thershold specified in \code{minPercentage}.
When is \code{FALSE}, the pie graph uses all the species catches which are greather than zero.}

\item{minPercentage}{Numeric value indicating the minimum percentage that catches must have
to be graphic. By default is 0.2.}

\item{...}{Extra arguments.}
}
\value{
A pie graph.
}
\description{
This method takes a \code{bitacora} object and do a pie graph.
}

