% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imarpe-main.R
\name{plotSpeciesComposition}
\alias{plotSpeciesComposition}
\title{PlotSpeciesComposition method}
\usage{
plotSpeciesComposition(x, ...)
}
\arguments{
\item{x}{Object of \code{bitacora} class.}

\item{...}{Extra arguments passed to \code{plotSpeciesComposition.bitacora} function.}
}
\description{
Method for plotSpeciesComposition.bitacora function.
}
\details{
For more details read the help of \code{\link{plotSpeciesComposition.bitacora}}.
}
\author{
Criscely Lujan-Paredes, \email{criscelylujan@gmail.com}.
}
