% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitacora-class.R
\name{print.bitacora}
\alias{print.bitacora}
\title{Print method for bitacora objects}
\usage{
\method{print}{bitacora}(x, language = "spanish")
}
\arguments{
\item{x}{Object of class \code{bitacora}.}

\item{language}{The select language to print the outputs.}
}
\description{
Shows main information from bitacora objects, like the origen of the
information, the number of trips, the number of data ports, the number of months
of data, the number of years of data, the types of fleets on the data.
}
\examples{
# Read a example of data base
bitacoraData = system.file("extdata", "bitacoraData.csv", package = "imarpe")

# Create a object of bitacora class
bitacoraObject = getBitacoraData(file = bitacoraData)
class(bitacoraObject)

# Show main information of the bitacora class object
print(bitacoraObject)
print(bitacoraObject, language = "english")

}

