% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitacoras-class.R
\name{print.bitacoras}
\alias{print.bitacoras}
\title{Print General Description of Bitacoras Objects}
\usage{
\method{print}{bitacoras}(x, ...)
}
\arguments{
\item{x}{An object of the class \dQuote{bitacoras}.}

\item{\dots}{Further arguments accepted by \code{\link{print}}.}
}
\value{
None (invisible NULL).
}
\description{
Prints the main features of the \dQuote{bitacoras} database such as file name, number of records, observers, ports
and involved years, and returns them invisibly (via \code{\link{invisible}}).
}
\details{
It is not useful to print all the content of large databases as \dQuote{bitacoras} object.
For this reason, this function is useful to print only metadata of \dQuote{bitacoras} object.
}
\examples{
## Loading the database of the class 'bitacoras'
data(Bitacoras)

## Verifying the class of the data
class(Bitacoras)

## S3 method for class 'bitacoras'
print(Bitacoras)
}
\author{
Erick Chacon Montalvan, Vilma Romero Romero, Criscely Lujan Paredes, Ricardo Oliveros-Ramos.

Maintainer: Ricardo Oliveros-Ramos <\email{roliveros@imarpe.gob.pe}>
}
\references{
Lujan Paredes C., Oliveros-Ramos R., Chacon Montalvan E., Romero Romero V., 2014.
Introduction to \pkg{imarpe} package for the automation of graphs, charts and reports using \code{R}.
}
\seealso{
\code{\link{getData}}, \code{\link{plot.bitacoras}}, \code{\link{summary.bitacoras}}
}

