% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery-class.R
\name{print.fishery}
\alias{print.fishery}
\title{Print method for fishery objects}
\usage{
\method{print}{fishery}(x, language = "spanish")
}
\arguments{
\item{x}{Object of class \code{fishery}.}

\item{language}{The select language to print the outputs.}
}
\description{
Shows main information from fishery objects, like number of records data,
the time period of date (years and months), the number of ports, the analyzed species and
the analyzed variable type.
}

