% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landings-class.R
\name{print.landings}
\alias{print.landings}
\title{Print General Description of Landings Objects}
\usage{
\method{print}{landings}(x, ...)
}
\arguments{
\item{x}{An object of the class \dQuote{landings}.}

\item{\dots}{Further arguments accepted by \code{\link{print}}}
}
\value{
None (invisible NULL).
}
\description{
Prints the main features of the \dQuote{landings} database such as the number of records, months, years and ports,
and returns them invisibly (via \code{\link{invisible}}).
}
\details{
It is not useful to print all the content of large databases as \dQuote{landings} objects.
For this reason, this function is useful to print only metadata of \dQuote{landings} objects.
}
\examples{
## Loading the database of the class 'landings'
data(Landings)

## Verifying the class of data
class(Landings)

## S3 method for class 'landings'
print(Landings)
}
\author{
Erick Chacon Montalvan, Vilma Romero Romero, Criscely Lujan Paredes, Ricardo Oliveros-Ramos.

Maintainer: Ricardo Oliveros-Ramos <\email{roliveros@imarpe.gob.pe}>
}
\references{
Lujan Paredes C., Oliveros-Ramos R., Chacon Montalvan E., Romero Romero V., 2014.
Introduction to \pkg{imarpe} package for the automation of graphs, charts and reports using \code{R}.
}
\seealso{
\code{\link{getData}}, \code{\link{plot.landings}}, \code{\link{summary.landings}}
}

