% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitacora-class.R
\name{print.summary.bitacora}
\alias{print.summary.bitacora}
\title{Print method for summary.bitacora}
\usage{
\method{print}{summary.bitacora}(x, language = "spanish")
}
\arguments{
\item{x}{Object of class \code{summary.bitacora}.}

\item{language}{The select language to print the summary of bitacora objects.}
}
\value{
Each element of \code{summary.bitacora} method.
}
\description{
Shows main information from \code{summary.bitacora} objects.
}
\examples{
# Read a example of data base
bitacoraData = system.file("extdata", "bitacoraData.csv", package = "imarpe")

# Create a object of bitacora class
bitacoraObject = getBitacoraData(file = bitacoraData)
class(bitacoraObject)

# Print the summary of the fishery class object
sumBitacora = summary(bitacoraObject)
print(sumBitacora)

sumBitacora = summary(bitacoraObject, language = "english")
print(sumBitacora, language = "english")

}
