% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landings-class.R
\name{print.summary.landings}
\alias{print.summary.landings}
\title{Print General Description of Summary.landings Objects}
\usage{
\method{print}{summary.landings}(x, ...)
}
\arguments{
\item{x}{An object of class \dQuote{summary.landings}.}

\item{\dots}{Further arguments accepted by \code{\link{print}}}
}
\value{
None (invisible NULL).
}
\description{
Prints summary information about a \dQuote{landings} object, which includes  the metadata, landings by ports, monthly landings and annual landings.
}
\details{
\tabular{ll}{
\code{info} \tab General overview of the landings database (number of records, months, years and ports). \cr
\code{ports} \tab Table indicating the number of landings by port.\cr
\code{months} \tab Table indicating the number of landings by month.\cr
\code{years} \tab Table indicating the number of landings by year.
}
}
\examples{
## Loading the database of the class 'landings'
data(Landings)

## Printing the summary
summary(Landings)
}
\author{
Erick Chacon Montalvan, Vilma Romero Romero, Criscely Lujan Paredes, Ricardo Oliveros-Ramos.

Maintainer: Ricardo Oliveros-Ramos <\email{roliveros@imarpe.gob.pe}>
}
\references{
Lujan Paredes C., Oliveros-Ramos R., Chacon Montalvan E., Romero Romero V., 2014.
Introduction to \pkg{imarpe} package for the automation of graphs, charts and reports using \code{R}.
}
\seealso{
\code{\link{getData}}, \code{\link{summary.landings}}, \code{\link{print.landings}}, \code{\link{plot.landings}}
}

