% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survey-class.R
\name{print.survey}
\alias{print.survey}
\title{Print General Description of Survey Objects}
\usage{
\method{print}{survey}(x, ...)
}
\arguments{
\item{x}{An object of the class \dQuote{survey}.}

\item{\dots}{Further arguments accepted by \code{\link{print}}.}
}
\value{
None (invisible NULL).
}
\description{
Prints the main features of the \dQuote{survey} database such as file name, dates, trawl numbers, latitudes
and year, and returns them invisibly (via \code{\link{invisible}}).
}
\details{
It is not useful to print all the content of large databases as \dQuote{survey} object.
For this reason, this function is useful to print only metadata of \dQuote{survey} object.
}
\examples{
## Loading the database of the class 'survey'
data(Survey)

## Verifying the class of the data
class(Survey)

## S3 method for class 'survey'
print(Survey)
}
\author{
Giancarlo Moron, Ricardo Oliveros-Ramos.

Maintainer: Ricardo Oliveros-Ramos <\email{roliveros@imarpe.gob.pe}>
}
\references{
Lujan Paredes C., Oliveros-Ramos R., Chacon Montalvan E., Romero Romero V., 2014.
Introduction to \pkg{imarpe} package for the automation of graphs, charts and reports using \code{R}.
}
\seealso{
\code{\link{getData}}, \code{\link{plot.survey}}, \code{\link{summary.survey}}
}

