% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitacora-class.R
\name{report.bitacora}
\alias{report.bitacora}
\title{Report method for bitacora objects}
\usage{
\method{report}{bitacora}(x, format = "latex", tangle = FALSE,
  output = NULL, open = TRUE)
}
\arguments{
\item{x}{Object of \code{bitacora} class.}
}
\description{
Export a report of bitacora.
}

