% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-class.R
\name{report.cpue}
\alias{report.cpue}
\title{Report method for cpue objects}
\usage{
\method{report}{cpue}(x, format = "latex", tangle = FALSE, output = NULL,
  open = TRUE, daysToPlot = c(1, 8, 15, 22), textAxis2 = NULL,
  textAxis4 = NULL)
}
\arguments{
\item{x}{Object of \code{cpue} class.}

\item{daysToPlot}{If is a daily plot by default the x axis plot the first day of the month
(1, 8, 15, 22). This is including in a vector form.}

\item{textAxis2}{The text of the x axis.}

\item{textAxis4}{The text of the y axis.}
}
\description{
Export a report of catch per unit effort class.
}

