% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery-class.R
\name{report.fishery}
\alias{report.fishery}
\title{Report method for fishery objects}
\usage{
\method{report}{fishery}(x, type = "singled", format = "latex",
  tangle = FALSE, output = NULL, open = TRUE, daysToPlot = c(1, 8, 15,
  22), textAxis2 = NULL, textAxis4 = NULL)
}
\arguments{
\item{x}{Object of \code{fishery} class.}

\item{type}{Complete this.}

\item{daysToPlot}{If is a daily plot by default the x axis plot the first day of the month
(1, 8, 15, 22). This is including in a vector form.}

\item{textAxis2}{The text of the x axis.}

\item{textAxis4}{The text of the y axis.}
}
\description{
Export a report of landing or fishing effort.
}

