% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery-class.R
\name{report}
\alias{report}
\title{Report method for fishery objects.}
\usage{
\method{report}{fishery}(x, ...)
}
\arguments{
\item{x}{Object of class \code{fishery}.}
}
\description{
Export a report of landing or fishing effort in a pdf document.
}

