% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imarpe-data.R
\docType{data}
\name{seasonAnchovyNC}
\alias{seasonAnchovyNC}
\title{Anchovy season for the north-central region.}
\format{A \code{data.frame} with four columns: year, the start date of the season for the
year, the end date of the season, and a code for the season.}
\usage{
seasonAnchovyNC
}
\description{
Historical records of the opening and closing of anchovy fishing
seasons in the north-central region of the Peruvian coast.
}
\references{
imarpe package vignette.
}
