% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitacora-class.R
\name{summary.bitacora}
\alias{summary.bitacora}
\title{Summary method for bitacora objects}
\usage{
\method{summary}{bitacora}(object, language = "spanish", latByPort = FALSE)
}
\arguments{
\item{object}{Object of class \code{bitacora}.}

\item{language}{The select language to print the summary of bitacora objects.}

\item{latByPort}{\code{logical}. By default (\code{FALSE}) to get the
latitude accoding the log book information, and to get the latitude
according the port (\code{TRUE}).}
}
\value{
A \code{list} of summary.bitacora class. This contains:
\itemize{
  \item observedTrip The observed trips by port.
  \item fishingHaul The fishing haul by latitude.
}
}
\description{
Get summary information of bitacora object.
}
\examples{
# Read a example of data base
bitacoraData = system.file("extdata", "bitacoraData.csv", package = "imarpe")

# Create a object of bitacora class
bitacoraObject = getBitacoraData(file = bitacoraData)
class(bitacoraObject)

# Produce the summary of bitacora class object
summary(bitacoraObject)

}

