% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.bitacoras}
\alias{summary.bitacoras}
\title{Create a Summary.bitacoras Object}
\usage{
\method{summary}{bitacoras}(object, ...)
}
\arguments{
\item{object}{An object of class \dQuote{bitacoras}.}

\item{\dots}{Further arguments.}
}
\value{
Returns an object of class \dQuote{summary.bitacoras} containing the following features:
\tabular{ll}{
\code{info} \tab General overview, metadata, of the \dQuote{bitacoras} database (number of records, observers, ports and involved years).\cr
\code{composition} \tab Dataframe indicating the number of captures by fish species.\cr
\code{effort} \tab Dataframe indicating the travel duration and the number of coves by port.\cr
\code{observer} \tab Dataframe indicating the number of observers by port.\cr
\code{set} \tab Dataframe indicating the number of coves by latitude.\cr
\code{depth} \tab Dataframe indicating the upper limit, lower limit and average depth by latitude.
}
}
\description{
Creates a new object of class \dQuote{summary.bitacoras}.
This class has summary information about a \dQuote{bitacoras} object, which includes the metadata, captures by specie,
travel duration and number of coves, observers by port, coves by latitude and depth by latitude.
}
\examples{
## Loading the database of the class 'bitacoras'
data(Bitacoras)

## Verifying the class of the data
class(Bitacoras)

## S3 method for class 'bitacoras'
SummaryBitacoras = summary(Bitacoras)
}
\author{
Erick Chacon Montalvan, Vilma Romero Romero, Criscely Lujan Paredes, Ricardo Oliveros-Ramos.

Maintainer: Ricardo Oliveros-Ramos <\email{roliveros@imarpe.gob.pe}>
}
\references{
Lujan Paredes C., Oliveros-Ramos R., Chacon Montalvan E., Romero Romero V., 2014.
Introduction to \pkg{imarpe} package for the automation of graphs, charts and reports using \code{R}.
}
\seealso{
\code{\link{getData}}, \code{\link{plot.bitacoras}}, \code{\link{print.bitacoras}}
}

