% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-class.R
\name{summary.cpue}
\alias{summary.cpue}
\title{Summary method for cpue objects}
\usage{
\method{summary}{cpue}(object, language = "spanish")
}
\arguments{
\item{object}{Object of \code{cpue} class.}

\item{language}{The select language to print the summary of cpue objects.
It could be \code{"spanish"} by default or \code{"english"}.}
}
\value{
A \code{list} of summary.cpue class. This contains:
\itemize{
  \item effort The type of the effort type that has been to analyze.
  \item portDay A data frame with the cpue by day and port.
  \item day A data frame with the cpue by day.
  \item port A data frame with the cpue by ports.
  \item months A data frame with the cpue by months.
  \item years A data frame with the cpue by years.
}
}
\description{
Get summary information included on objects of catch
per unit effort (cpue) class.
}
\examples{
# Read a example of a data base
fisheryData = system.file("extdata", "fisheryData.csv", package = "imarpe")

# Produce a object of cpue class
cpue  = getFishingData(file = fisheryData, type = "cpue", varType = "cpue", sp = "caballa", efforType = "capacidad_bodega")
class(cpue)

# Produce the summary of the cpue class object
summary(cpue)

}
