% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landings-class.R
\name{summary.landings}
\alias{summary.landings}
\title{Create a Summary.landings Object}
\usage{
\method{summary}{landings}(object, ...)
}
\arguments{
\item{object}{An object of class \dQuote{landings}.}

\item{\dots}{Further arguments}
}
\value{
Returns an object of class \dQuote{summary.landings} containing the following features:
\tabular{ll}{
\code{info} \tab General overview, metadata, of the landings database (number of records, months, years and ports).\cr
\code{sumPorts} \tab Dataframe containing the number of ports by day.\cr
\code{ports} \tab Dataframe indicating the number of landings by port.\cr
\code{months} \tab Dataframe indicating the number of landings by month.\cr
\code{years} \tab Dataframe indicating the number of landings by year.
}
}
\description{
Creates a new object of class \dQuote{summary.landings}. This class has summary information about a \dQuote{landings} object, which includes the metadata,
captures by specie, number of ports by day, landings by port, landings by month and landings by year.
}
\examples{
## Loading the database of the class 'landings'
data(Landings)

## Verifying the class of the data
class(Landings)

## S3 method for class 'landings'
SummaryLandings = summary(Landings)
}
\author{
Erick Chacon Montalvan, Vilma Romero Romero, Criscely Lujan Paredes, Ricardo Oliveros-Ramos.

Maintainer: Ricardo Oliveros-Ramos <\email{roliveros@imarpe.gob.pe}>
}
\references{
Lujan Paredes C., Oliveros-Ramos R., Chacon Montalvan E., Romero Romero V., 2014.
Introduction to \pkg{imarpe} package for the automation of graphs, charts and reports using \code{R}.
}
\seealso{
\code{\link{getData}}, \code{\link{plot.landings}}, \code{\link{summary.landings}}
}

