% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survey-class.R
\name{summary.survey}
\alias{summary.survey}
\title{Create a Summary.survey Object}
\usage{
\method{summary}{survey}(object, ...)
}
\arguments{
\item{object}{An object of class \dQuote{survey}.}

\item{\dots}{Further arguments.}
}
\value{
Returns an object of class \dQuote{summary.survey} containing the following features:
\tabular{ll}{
\code{info} \tab General overview, metadata, of the \dQuote{survey} database (year, survey name, start and end of the survey).\cr
\code{boats} \tab Dataframe indicating trawls and latitudes by boat.\cr
\code{trawl} \tab Dataframe indicating the number of trawls by latitude.\cr
\code{composition} \tab Dataframe indicating the most abundant species.\cr
\code{diversity} \tab Dataframe indicating four different diversity indices.
}
}
\description{
Creates a new object of class \dQuote{summary.survey}.
This class has summary information about a \dQuote{survey} object, which includes the metadata, Boat names,
trawls by latitude, sample composition, diversity.
}
\examples{
## Loading the database of the class 'survey'
data(Survey)

## Verifying the class of the data
class(Survey)

## S3 method for class 'survey'
SummarySurvey = summary(Survey)
}
\author{
Giancarlo Moron, Ricardo Oliveros-Ramos.

Maintainer: Ricardo Oliveros-Ramos <\email{roliveros@imarpe.gob.pe}>
}
\references{
Lujan Paredes C., Oliveros-Ramos R., Chacon Montalvan E., Romero Romero V., 2014.
Introduction to \pkg{imarpe} package for the automation of graphs, charts and reports using \code{R}.
}
\seealso{
\code{\link{getData}}, \code{\link{plot.survey}}, \code{\link{print.survey}}
}

