% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{vectorColours}
\alias{vectorColours}
\title{Vector colours}
\usage{
vectorColours(n)
}
\arguments{
\item{n}{Number of color levels.}
}
\value{
A vector giving the code of colours.
}
\description{
Function to use an additional vector of colours. This function use the
the function \code{tim.colors} of the fields packages.
}
\examples{
vectorColours(n = 5)

}

