% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{word2month}
\alias{word2month}
\title{Months in words to months in numbers}
\usage{
word2month(x)
}
\arguments{
\item{x}{A vector with the abbreviation of months in words.}
}
\value{
A vector with the months in numbers. Where January is number one
and increasing until December that is the number twelve.
}
\description{
Function to convert abbreviations of months in words to months in number.
}
\examples{
# The 'months' vector have the twelve abbreviation months
months = month.abb
word2month(x = months)

}

