% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery-class.R
\name{combineFisheryVar}
\alias{combineFisheryVar}
\title{Combine fishery variables}
\usage{
combineFisheryVar(landing, effort)
}
\arguments{
\item{landing}{List of landings with the class fishery.}

\item{effort}{List of efforts with the class fishery.}
}
\value{
A list of the class fishery with information about the two lists.
}
\description{
Function to combine two fishery class outputs (landings and effort)
into a only list.
}

