% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery-class.R
\name{combineFisheryVar}
\alias{combineFisheryVar}
\title{Combine fishery variables}
\usage{
combineFisheryVar(landing, effort)
}
\arguments{
\item{landing}{Object of \code{fishery} class and landing variable type.}

\item{effort}{Object of \code{fishery} class and effort variable type.}
}
\value{
A object of \code{fishery} class with information about the two lists, the first
with for landing and the second for effort.
}
\description{
Function to combine two fishery class outputs (landing and effort)
into a only list.
}
\examples{
# Read a example of a data base
fisheryData = system.file("extdata", "fisheryData.csv", package = "imarpe")

# A fishery class object with a landing variable type
landingObject = getFishingData(file = fisheryData, type = "fisheryinfo", varType = "landing",
 sp = "caballa")

# A fishery class object with a effort variable type
effortObject = getFishingData(file = fisheryData, type = "fisheryinfo", varType = "effort",
sp = "caballa", efforType = "viaje")

# Use combineFisheryVar
fisheryVar = combineFisheryVar(landing = landingObject, effort = effortObject)
class(fisheryVar)
}

