% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getBitacoraData}
\alias{getBitacoraData}
\title{Title}
\usage{
getBitacoraData(file, colTrip = "CODIGO_VIAJE", colPort = "PUERTO_SALIDA",
  colDateStart = "DIA_SALIDA", colDateOut = "DIA_ARRIBO",
  colSearchTime = "DURACION_BUSQUEDA",
  colStorageCapacity = "CAPACIDAD_BODEGA_REGISTRADA",
  colLat = "LATITUD_INICIAL", colLon = "LONGITUD_INICIAL",
  colHaul = "NUMERO_CALA", colHaulTotal = "TOTAL_CALAS",
  colCatchHaul = "CAPTURA_CALA", capAnch = "CAPTURA_ANCHOVETA",
  capSar = "CAPTURA_SARDINA", capJur = "CAPTURA_JUREL",
  capCab = "CAPTURA_CABALLA", capBon = "CAPTURA_BONITO")
}
\arguments{
\item{capBon}{}
}
\description{
Title
}

