% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imarpe-main.R
\name{getBitacoraData}
\alias{getBitacoraData}
\title{Get data from bitacora class}
\usage{
getBitacoraData(file, colTrip = "CODIGO_VIAJE", colPort = "PUERTO_SALIDA",
  colDateStart = "DIA_SALIDA", colDateOut = "DIA_ARRIBO",
  colSearchTime = "DURACION_BUSQUEDA",
  colStorageCapacity = "CAPACIDAD_BODEGA_REGISTRADA",
  colLat = "LATITUD_INICIAL", colLon = "LONGITUD_INICIAL",
  colHaul = "NUMERO_CALA", colHaulTotal = "TOTAL_CALAS",
  colCatchHaul = "CAPTURA_CALA", capAnch = "CAPTURA_ANCHOVETA",
  capSar = "CAPTURA_SARDINA", capJur = "CAPTURA_JUREL",
  capCab = "CAPTURA_CABALLA", capBon = "CAPTURA_BONITO")
}
\arguments{
\item{file}{The name of the file. If it does not contain an absolute path, the file name is relative to
the current working directory, \code{\link{getwd}}.}

\item{colTrip}{Name of travel code column. By default is "CODIGO_VIAJE".}

\item{colPort}{Name of the port column. By default is "PUERTO_SALIDA".}

\item{colDateStart}{Name of data column with the start date of the travel. By default is "DIA_SALIDA".}

\item{colDateOut}{Name of data column with the end date of the travel. By default is "DIA_ARRIBO".}

\item{colSearchTime}{Name of the data column with the search time. By default is "DURACION_BUSQUEDA".}

\item{colStorageCapacity}{Name of the data column with the storage capacity. By default is "CAPACIDAD_BODEGA_REGISTRADA".}

\item{colLat}{Name of the data column with the latitude. By default is "LATITUD_INICIAL".}

\item{colLon}{Name of the data column with the longitude. By default is "LONGITUD_INICIAL".}

\item{colHaul}{Name of the data column with the fishing haul number. By default is "NUMERO_CALA".}

\item{colHaulTotal}{Name of the data column with the total number of fishing hauls by travel. By default is "TOTAL_CALAS".}

\item{colCatchHaul}{Name of the data column with the catch by fishing haul. By default is "CAPTURA_CALA".}

\item{capAnch}{Name of the data column with the anchovy catches. By default is "CAPTURA_ANCHOVETA".}

\item{capSar}{Name of the data column with the sardine catches. By default is "CAPTURA_SARDINA".}

\item{capJur}{Name of the data column with the jack mackerel catches. By default is "CAPTURA_JUREL".}

\item{capCab}{Name of the data column with the chub mackerel catches. By default is "CAPTURA_CABALLA".}

\item{capBon}{Name of the data column with the bonito catches. By default is "CAPTURA_BONITO".}
}
\value{
\tabular{ll}{
\code{data} \tab A data frame containing a representation of the data in the file.\cr
\code{info} \tab A list containing the main features of the data.
}
}
\description{
This function read the information from the Programme of Fishery Logbooks (in spanish: Programa de
Bitacoras de Pesca - PBP) and create a data base which includes a data frame and a list with the main features
of the data frame. The included data frame object has observations corresponding to rows and variables to columns.
}
\note{
If type is not explicitly defined, the function will be equivalent to use \code{\link{read.csv}}.
}
\examples{
# Read a example of data base
bitacoraData = system.file("extdata", "bitacoraData.csv", package = "imarpe")

# Create a object of bitacora class
bitacoraObject = getBitacoraData(file = bitacoraData)

# Check the class of bitacoraObject
class(bitacoraObject)

# To get the data of bitacora class
dataBase = bitacoraObject$data

# To get the main features of the data
info = bitacoraObject$info

}
\author{
Criscely Lujan-Paredes, \email{criscelylujan@gmail.com}.
}
