% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imarpe-main.R
\name{getDailyReport}
\alias{getDailyReport}
\title{Get daily report of fishing monitoring}
\usage{
getDailyReport(directory = NULL, datesList, simpleFreqSizes, dataCruise,
  dataCruiseCsv, officialBiomass = NULL, readLocalPorcentas = FALSE,
  dirPorcentas = NULL, porcentaPrefix = "/imarpe_rpelag_porfinal",
  addEnmalle = TRUE, enmalleParams = list(mean = 11, sd = 5.5, maxProportion
  = 0.2 * 0.05), prop_enmalleParams = list(peak = 7, top = 11, asc_width =
  exp(1), dsc_width = exp(1), init = 0, final = 0),
  urlFishingMonitoring = "http://www.imarpe.pe/imarpe/archivos/reportes/imarpe_rpelag_porfinal",
  threshold = 30, species = "Anchoveta", a = NULL, b = NULL,
  growthParameters = list(k = 0.83, Linf = 19.21, sizeM = c(0, 8, 12), vectorM
  = c(1.29, 0.92, 0.83), catchFactor = 1, scenario = "neutro"))
}
\arguments{
\item{directory}{Directory where the anchovy landing reports are stored. By default it
is \code{NULL}, temporarily saved and then deleted. If you want to keep this parameter
must be changed.}

\item{datesList}{A \code{list} of six dates (format: YEAR-month-day) with the names:
 \itemize{
  \item surveyDate: finish date of reference survey.
  \item startDate: start date to download the landing files.
  \item endDate: end date to download the landing files.
  \item endSeasonDate: end date of fishing season.
}}

\item{simpleFreqSizes}{A comma delimited file (.csv) with simple frequency data per size.}

\item{dataCruise}{A RData with the outputs of the cruise.}

\item{officialBiomass}{A official value of biomass.}

\item{addEnmalle}{\code{logical} which indicates whether to add or not (default) 'enmalle' influence to catches.}

\item{enmalleParams}{If \code{addEnmalle = TRUE}, a \code{list} with main parameters to describe 'enmalle':
 \itemize{
  \item mean: the meanfor 'enmallados' length.
  \item sd: the SD for 'enmallados' length.
  \item maxProportion: factor of 'enmallados'.
}}

\item{urlFishingMonitoring}{The web address (url - Uniform Resource Locator) for downloading
the landings. By default it is \url{http://www.imarpe.pe/imarpe/archivos/reportes/imarpe_rpelag_porfinal}.}

\item{threshold}{Threshold for considering the number of individuals. By default \code{threshold = 30}.}

\item{species}{The species that is going to be analyzed. By default is \code{species = "Anchoveta"}.}

\item{a}{Length-weight ratio parameter. By default, it is \code{NULL}, check Details.}

\item{b}{Length-weight ratio parameter. By default, it is \code{NULL}, check Details.}

\item{growthParameters}{A \code{list} of growthParamters. By default this are:
\itemize{
 \item k = 0.83
 \item Linf = 19.21
 \item sizeM = c(0, 8, 12)
 \item vectorM = rep(0.8, 3)
 \item catchFactor = 1
 \item scenario = "neutro"
}}
}
\value{
A object of fishingMonitoring class. It is saved on the working directory.
}
\description{
This function download anchovy landing information from an official
repository of the IMARPE to compare this landings with the biomass estimated.
}
\details{
Allometric growth parameters \code{a} and \code{b} are completely necesary for calculations, if they
are \code{NULL} (default), the users must be sure that the \code{dataCruise} file has them inside (that is true if 
\code{dataCruise} comes from \code{TBE} package).
}
\author{
Wencheng Lau-Medrano, \email{luis.laum@gmail.com}, Josymar Torrejon and Pablo Marin.
}
