% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getDailyReport}
\alias{getDailyReport}
\title{Title}
\usage{
getDailyReport(directorio, URLporcentas, fechaInicio, fechaFinal,
  fechaInicio_exploratoria, fechaFinal_exploratoria, axisParamsCuota,
  tallasSimple, datosCrucero, umbral, especie, a, b, officialBiomass, k, Linf,
  sizeM, vectorM, catchFactor)
}
\arguments{
\item{catchFactor}{}
}
\description{
Title
}

