% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getDailyReport}
\alias{getDailyReport}
\title{Get daily report of fishing monitoring}
\usage{
getDailyReport(directory = NULL,
  urlFishingMonitoring = "http://www.imarpe.pe/imarpe/archivos/reportes/imarpe_rpelag_porfinal",
  datesList, simpleFreqSizes, dataCruise, officialBiomass, threshold = 30,
  species = "Anchoveta", a, b, growthParameters = list(k = 0.83, Linf =
  19.21, sizeM = c(0, 8, 12), vectorM = rep(0.8, 3), catchFactor = 1))
}
\arguments{
\item{directory}{Directory where the anchovy landing reports are stored. By default it
is \code{NULL}, temporarily saved and then deleted. If you want to keep this parameter
must be changed.}

\item{urlFishingMonitoring}{The web address (url - Uniform Resource Locator) for downloading
the landings. By default it is \url{http://www.imarpe.pe/imarpe/archivos/reportes/imarpe_rpelag_porfinal}.}

\item{datesList}{A \code{list} of six dates (\code{Y-m-d}) with the names:
 \itemize{
  \item startDate: start date to download the landing files.
  \item endDate: end date to download the landing files.
  \item startExploringDate: start date of exploratory analysis.
  \item endExploringDate: end date of exploratory analysis.
  \item startSeasonDate: start date of fishing season.
  \item endSeasonDate: end date of fishing season.
}}

\item{simpleFreqSizes}{A comma delimited file (.csv) with simple frequency data per size.}

\item{dataCruise}{A RData with the outputs of the cruise.}

\item{officialBiomass}{A official value of biomass.}

\item{threshold}{Threshold for considering the number of individuals. By default \code{threshold = 30}.}

\item{species}{The species that is going to be analyzed. By default is \code{species = "Anchoveta"}.}

\item{a}{Length-weight ratio parameter.}

\item{b}{Length-weight ratio parameter.}

\item{growthParameters}{A \code{list} of growthParamters. By default this are:
\itemize{
 \item k = 0.83
 \item Linf = 19.21
 \item sizeM = c(0, 8, 12)
 \item vectorM = rep(0.8, 3)
 \item catchFactor = 1
}}
}
\value{
A object of fishingMonitoring class. It is saved on the working directory.
}
\description{
This function download anchovy landing information from an official
repository of the IMARPE to compare this landings with the biomass estimated.
}
\author{
Wencheng Lau-Medrano, \email{luis.laum@gmail.com}, Josymar Torrejon and Pablo Marin.
}

