% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getFishingData}
\alias{getFishingData}
\title{Get data from fishery class}
\usage{
getFishingData(file, type, varType, toTons = TRUE, sp, start = NULL,
  end = NULL, port = NULL, efforType = "viaje", fleeType = NULL, ...)
}
\arguments{
\item{file}{The name of a file with a comma separated values format (extension of this files are .csv).
This file have fishing information related to the landing and effort. If it does not contain an absolute path,
the file name is relative to the current working directory, \code{\link{getwd}}.}

\item{type}{The type of the information: fisheryInfo with two variables (landing and fishing effort) and
the cpue (catch per unith effort).}

\item{varType}{The type of the variable that has been to analyze. This can be: lading, effort and cpue.}

\item{toTons}{A logical parameter. \code{TRUE} (dafault) it assume that the fishery information is
in kilograms and converts it into tonnes (divided by 1000). \code{FALSE} it assume the the information
is in tons and don't convert the data.}

\item{sp}{The name of the species.}

\item{start}{The date to start the analysis. By dafault is \code{NULL}.}

\item{end}{The date to end the analysis. By dafault is \code{NULL}.}

\item{port}{The specific port to analyze. By dafault is \code{NULL} and analyze all ports.}

\item{efforType}{If the varType is effort or cpue (catch per unit effort) this parameter indicate what
type of effort will be used.
\itemize{
  \item "viaje" for the travel number.
  \item "capacidad_bodega" for storage capacity.
  \item "anzuelos" for the hook number.
  \item "embarcaciones" for the number of boats.
}}

\item{fleeType}{The fleet type to analyze. By default is \code{NULL}.}

\item{...}{Further arguments passed to \code{.fleetData} function.}
}
\value{
\tabular{ll}{
\code{data} \tab A data frame containing a representation of the data in the file.\cr
\code{info} \tab A list containing the main features of the data. \cr
\code{fleeTable} \tab A data frame with the fishery information by years and months by
each type of fleet.
}
}
\description{
This function read the fishing information (landings, effort and catch per unit effor) to
 create a data base which includes a data frame and a list with the main features of the data frame.
 The included data frame object has observations corresponding to rows and variables to columns.
}
\note{
If type is not explicitly defined, the function will be equivalent to use \code{\link{read.csv}}.
}
\examples{
# Read a example of a data base
fisheryData = system.file("extdata", "fisheryData.csv", package = "imarpe")

# For landing information assuming the analyzed species is 'caballa', put the name of the species on \\code{sp}
landing  = getFishingData(file = fisheryData, type = "fisheryinfo", varType = "landing", sp = "caballa")

# If you can analyzed a period of time you have to use \\code{start} and \\code{end} parameters
landing  = getFishingData(file = fisheryData, type = "fisheryinfo", varType = "landing", sp = "caballa",
start = "2009-04-10", end = "2009-08-30")

# If you can analyzed a specific port put the name of the port on \\code{port}
landing  = getFishingData(file = fisheryData, type = "fisheryinfo", varType = "landing", sp = "caballa",
start = "2009-04-10", end = "2009-08-30", port = "PAITA")

# Check the class of the landing object, it would be 'fishery'
class(landing)

# To get the data of the landing object.
dataBase = landing$data

# To get the main features of the landing object.
info = landing$info

# To get the data frame with the landings by each type of fleet.
fleet = landing$fleeTable


# To analyze the effort information on the same data, use the parameter varType and
# specify the type of effort in the parameter efforType
effort  = getFishingData(file = fisheryData, type = "fisheryinfo", varType = "effort",
 sp = "caballa", efforType = "capacidad_bodega")

# To analyze the cpue information on the same data change the parameter type, specify
the varType and specify the type of effort to calculate the catch per unit effort (cpue)
cpue  = getFishingData(file = fisheryData, type = "cpue", varType = "cpue",
sp = "caballa", efforType = "capacidad_bodega")

}
\author{
Criscely Lujan-Paredes, \email{criscelylujan@gmail.com}.
}
