% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getFishingData}
\alias{getFishingData}
\title{Obtener blabla}
\usage{
getFishingData(file, type, varType, toTons = TRUE, sp, start = NULL,
  end = NULL, port = NULL, efforType = "viaje", fleeType = NULL, ...)
}
\arguments{
\item{...}{}
}

