% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{imarpe}
\alias{imarpe}
\alias{imarpe-package}
\title{imarpe: Automatización de gráficas, tablas y reportes del Instituto del Mar del Peru (IMARPE).}
\description{
Este paquete proporciona herramientas para la elaboración de gráficas y tablas que se realizan de manera rutinaria en los informes producidos por el Instituto del Mar del Peru (IMARPE).
El paquete \pkg{imarpe} está diseñado para trabajar con funciones genéricas (e.g. \code{\link{plot}}, \code{\link{summary}}, \code{\link{print}}) sin importar el tipo de datos utilizados,
facilitando su uso y permitiendo su utilización incluso con conocimientos mínimos de \code{R}.
}
\details{
El entorno del paquete trabaja básicamente con cinco diferentes clases de objetos tales como: bitacoras, landing, length, composition and spatialDistribution.
}
\section{Información adicional}{

\tabular{ll}{
Package: \tab imarpe\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-03-11\cr
License: \tab GLP-2
}
}
\examples{
## Para la clase bitacoras
## Leer la data
data = Bitacoras
## Graficando el esfuerzo
plot(data,type="effort")
## Graficando la profundidad
plot(data,type="depth")

## Para la clase landings
## Leer la data
data = Landings
## Grafica por meses
plot(data,time="month")
## Grafica por años
plot(data,time="year")
}
\author{
Ricardo Oliveros-Ramos, Criscely Lujan Paredes, Erick Chacon Montalvan, Vilma Romero Romero.

Maintainer: Ricardo Oliveros-Ramos <\email{roliveros@imarpe.gob.pe}>
}
\references{
Lujan Paredes C., Oliveros-Ramos R., 2014.
Introducción al paquete \pkg{imarpe} para la automatización de gráficos, tablas y reportes usando \code{R}.
}
\seealso{
La siguiente lista muestra las funciones que se pueden utilizar para analizar los diferentes tipos de clases
Leer el conjunto de datos de un archivo con la función \code{\link{getData}}.
}
\concept{
IMARPE
}
\keyword{R}
\keyword{graphs}
\keyword{reports}
\keyword{tables}

