\name{imarpe-package}
\alias{imarpe-package}
\alias{imarpe}
\docType{package}
\title{
imarpe: automatizaci?n de gr?ficas, tablas y reportes del Instituto del Mar del Per?.
}
\description{
Este paquete proporciona herramientas para la elaboraci?n de gr?ficas y tablas que se realizan de manera rutinaria en los informes producidos por el Instituto del Mar del Per? (IMARPE). El paquete est? dise?ado para trabajar con funciones gen?ricas (e.g. \code{plot}, \code{summary}, \code{print}) sin importar el tipo de datos utilizado, facilitando su uso y permitiendo su utilizaci?n incluso con conocimientos m?nimos de \code{R}. 
}
\details{
\tabular{ll}{
Package: \tab imarpe\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-03-11\cr
License: \tab TBD\cr
}

}
\author{
Criscely Luj?n Paredes, Ricardo Oliveros-Ramos

Maintainer: Ricardo Oliveros-Ramos <roliveros@imarpe.gob.pe>
}
\references{
Luj?n-Paredes C, Oliveros-Ramos R. 2014. Introducci?n al paquete \code{imarpe} para la automatizaci?n de gr?ficos, tablas y reportes usando \code{R}.
}
\keyword{ figuras, tablas, reportes, imarpe, R. }
\seealso{
}
\examples{
~~ simple examples of the most important functions ~~
}
