% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-class.R
\name{plot.cpue}
\alias{plot.cpue}
\title{Plot method for cpue objects}
\usage{
\method{plot}{cpue}(x, language = "spanish", ploType = NULL,
  daysToPlot = c(1, 8, 15, 22), textAxis2 = NULL, textAxis4 = NULL,
  colBar = "gray", ...)
}
\arguments{
\item{x}{Object of \code{cpue} class.}

\item{language}{A \code{character}. Define the language of text labels in plots.
It could be \code{"spanish"} or \code{"english"}.}

\item{ploType}{What type of plot should be draw. Possible types are:
\itemize{
   \item plotDaily for daily plot
  \item plotMonthly for monthly plot
  \item plotYearly for yearly plot
  \item plotPeru for all the peruvian region
  \item plotNC to graph the north-central region
  \item plotS to graph the south region
}}

\item{daysToPlot}{If is a daily plot by default the x-axis show the first day of a week
(1, 8, 15, 22) but could be change to show a specific day or to show the all days of the
data with \code{all}. This is including in a vector form.}

\item{textAxis2}{The text of the x axis.}

\item{textAxis4}{The text of the y axis.}

\item{...}{Extra arguments.}
}
\value{
A graph of the specified type in \code{ploType}.
}
\description{
This method takes a \code{cpue} object and make useful plots.
The plots can be daily, monthly, yearly, over the peruvian region, and
for north-central and south peruvian region.
}
\examples{
# Read a example of a data base
fisheryData = system.file("extdata", "fisheryData.csv", package = "imarpe")

# Produce a object of cpue class
cpue  = getFishingData(file = fisheryData, type = "cpue", varType = "cpue", sp = "caballa", efforType = "capacidad_bodega")
class(cpue)

plot(cpue)
plot(cpue, daysToPlot = "1", colBar = "red")
plot(cpue, language= "english")
plot(cpue, ploType = "plotMonthly")
plot(cpue, ploType = "plotMonthly", language= "english", colBar = "green")
plot(cpue, ploType = "plotYearly")
plot(cpue, ploType = "plotYearly", colBar = "orange")
plot(cpue, ploType = "plotPERU")
plot(cpue, ploType = "plotNC")
plot(cpue, ploType = "plotS")

}
