% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery-class.R
\name{plot.fishery}
\alias{plot.fishery}
\title{Plot method for fishery}
\usage{
\method{plot}{fishery}(x, language, ploType = NULL, daysToPlot = c(1, 8, 15,
  22), textAxis2 = NULL, textAxis4 = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{fishing}.}

\item{language}{\code{character}. Define the language of text labels in plots.}

\item{ploType}{What type of plot should be draw. Possible types are:
\itemize{
  \item plotDaily for daily plot
  \item plotMonthly for monthly plot
  \item plotYearly for yearly plot
  \item plotNC to plot the north-central region
  \item plotS to plot the south region
}}

\item{daysToPlot}{If is a daily plot by default the x axis plot the first day of the month
(1, 8, 15, 22). This is including in a vector form.}

\item{textAxis2}{The text of the x axis.}

\item{textAxis4}{The text of the y axis.}

\item{...}{Extra arguments.}
}
\description{
This method takes a \code{fishery} object and make useful plots for
each variables (lading and fishing effort). The plots can be daily, monthly, yearly or
for north-central and south peruvian region.
}

