% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitacora-internal.R
\name{plotEffort.numeric}
\alias{plotEffort.numeric}
\title{Plot series of fishing effort by ports}
\usage{
\method{plotEffort}{numeric}(effort1, effort2, labAxis2, labAxis4,
  colBar = NULL, colLine = NULL, legend = NULL, ...)
}
\arguments{
\item{effort1}{A vector of the first serie of fishing effort.}

\item{effort2}{A vector of the second serie of fishing effort.}

\item{labAxis2}{Text to be included on the left vertical axis of the graph}

\item{labAxis4}{Text to be included on the right vertical axis of the graph}

\item{colBar}{The colour of the barplot. Sea \code{\link{barplot}}.}

\item{colLine}{The colour of the line on the graph.}

\item{legend}{The names to include on the legend of the graph.}

\item{...}{Extra arguments.}
}
\value{
A graph for two types of fishing effort by ports. One type is on the left
 vertical axis and the other on the right vertical axis, the ports are on the horizontal axis.
}
\description{
This method take a series of fishing effort by ports and produces a graph
that is useful for the PBP program.
}
\examples{
# Read a example of data base
bitacoraData = system.file("extdata", "bitacoraData.csv", package = "imarpe")

# Create a object of bitacora class
bitacoraObject = getBitacoraData(file = bitacoraData)

# Get the results
mainBitacoraData = getMainResults.bitacora(object = bitacoraObject, language = "spanish",
specie = "anchoveta", effortData = TRUE)

# Effort data
effortData = mainBitacoraData$effortData

# Filter data: anchovy data for anchovy and the north-central region
effortNC = effortSpeciesData.bitacora(data = effortData, species = "anchoveta",
 region = "norte-centro")

# Get effort
effortNC_tt = getEffort(data = effortNC, efforType = "travelTime", effortBy = "port")
effortNC_hn = getEffort(data = effortNC, efforType = "haulTotal", effortBy = "port")

# Graph
plotEffort(effort1 = effortNC_tt, effort2 = effortNC_hn, labAxis2 = "Horas de viaje",
labAxis4 = "N\\u{00FA}mero de calas", colBar="gray", colLine = "black",
legend = c("Duracion del viaje", "N\\u{00FA}mero de calas"))

}
\author{
Criscely Lujan-Paredes, \email{criscelylujan@gmail.com}.
}

