% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-class.R
\name{print.cpue}
\alias{print.cpue}
\title{Print method for cpue objects}
\usage{
\method{print}{cpue}(x, language = "spanish")
}
\arguments{
\item{x}{Object of class \code{cpue}.}

\item{language}{The select language to print the outputs.}
}
\description{
Shows main information from catch per unit effort (cpue) objects, like
number of records data, the time period of date (years and months), the number of
ports, the analyzed species and the type of effort used on the cpue.
}

