% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery-class.R
\name{print.fishery}
\alias{print.fishery}
\title{Print method for fishery objects}
\usage{
\method{print}{fishery}(x, language = "spanish")
}
\arguments{
\item{x}{Object of \code{fishery} class.}

\item{language}{The select language to print the outputs.}
}
\description{
Shows main information from objects of the \code{fishery} class,
like number of records data, the time period of dates (years and months),
the number of ports, the analyzed species and the analyzed variable type.
}
\examples{
# Read a example of a data base
fisheryData = system.file("extdata", "fisheryData.csv", package = "imarpe")

# Produce a object of fishery class
landing  = getFishingData(file = fisheryData, type = "fisheryinfo", varType = "landing", sp = "caballa")
class(landing)

# Show main information of the fishery class object
print(landing)
print(landing, language = "english")

}
