% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitacoras-class.R, R/survey-class.R
\name{print.summary.bitacoras}
\alias{print.summary.bitacoras}
\title{Print General Description of Summary.bitacoras Objects}
\usage{
\method{print}{summary.bitacoras}(x, ...)

\method{print}{summary.bitacoras}(x, ...)
}
\arguments{
\item{x}{An object of class \dQuote{summary.bitacoras}.}

\item{\dots}{Further arguments accepted by \code{\link{print}}}

\item{x}{An object of class \dQuote{summary.survey}.}

\item{\dots}{Further arguments accepted by \code{\link{print}}}
}
\value{
None (invisible NULL).

None (invisible NULL).
}
\description{
Prints summary information about a \dQuote{bitacoras} object, which includes  the metadata, captures by specie,
travel duration and number of coves, observers by port, coves by latitude and depth by latitude.

Prints summary information about a \dQuote{survey} object, which includes the metadata, Boat names,
trawls by latitude, sample composition, diversity.
}
\details{
\tabular{ll}{
\code{info} \tab General overview of the bitacoras database (number of records, observers, ports and involved years). \cr
\code{Species composition} \tab Table indicating the number and percentage of captures by fish species.\cr
\code{Sampling effort} \tab Table indicating the travel duration and the average number of coves by port.\cr
\code{Number of observers} \tab Table indicating the number of observers by port.\cr
\code{Number of sets by latitude} \tab Table indicating the number of coves by latitude.\cr
\code{Depth by latitude} \tab Table indicating the upper limit, lower limit and average depth by latitude.
}

\tabular{ll}{
\code{info} \tab General overview of the survey database (year, survey name, start and end of the survey). \cr
\code{Boats information} \tab Dataframe indicating trawls and latitudes by boat.\cr
\code{Trawls by latitude} \tab Table indicating the number of trawls by latitude.\cr
\code{Species composition} \tab Dataframe indicating the most abundant species.\cr
\code{Diversity} \tab Dataframe indicating four different diversity indices.
}
}
\examples{
## Loading the database of the class 'bitacoras'
data(Bitacoras)

## Printing the summary
summary(Bitacoras)
## Loading the database of the class 'bitacoras'
data(Survey)

## Printing the summary
summary(Survey)
}
\author{
Erick Chacon Montalvan, Vilma Romero Romero, Criscely Lujan Paredes, Ricardo Oliveros-Ramos.

Maintainer: Ricardo Oliveros-Ramos <\email{roliveros@imarpe.gob.pe}>

Giancarlo Moron, Ricardo Oliveros-Ramos.

Maintainer: Ricardo Oliveros-Ramos <\email{roliveros@imarpe.gob.pe}>
}
\references{
Lujan Paredes C., Oliveros-Ramos R., Chacon Montalvan E., Romero Romero V., 2014.
Introduction to \pkg{imarpe} package for the automation of graphs, charts and reports using \code{R}.

Lujan Paredes C., Oliveros-Ramos R., Chacon Montalvan E., Romero Romero V., 2014.
Introduction to \pkg{imarpe} package for the automation of graphs, charts and reports using \code{R}.
}
\seealso{
\code{\link{getData}}, \code{\link{summary.bitacoras}}, \code{\link{print.bitacoras}}, \code{\link{plot.bitacoras}}

\code{\link{getData}}, \code{\link{summary.bitacoras}}, \code{\link{print.bitacoras}}, \code{\link{plot.bitacoras}}
}

