% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-class.R
\name{print.summary.cpue}
\alias{print.summary.cpue}
\title{Print method for summary.cpue}
\usage{
\method{print}{summary.cpue}(x, language = "spanish")
}
\arguments{
\item{x}{Object of class \code{summary.cpue}.}

\item{language}{The select language to print the summary of cpue objects.
It could be \code{"spanish"} by default or \code{"english"}.}
}
\value{
Each element of \code{summary.cpue} method.
}
\description{
Shows main information from \code{summary.cpue} objects.
}

