% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery-class.R
\name{print.summary.fishery}
\alias{print.summary.fishery}
\title{Print method for summary.fishery}
\usage{
\method{print}{summary.fishery}(x, language = "spanish")
}
\arguments{
\item{x}{Object of class \code{summary.fishery}.}

\item{language}{The select language to print the summary of fishery objects.
It could be \code{"spanish"} by default or \code{"english"}.}
}
\value{
Each element of \code{summary.fishery} method.
}
\description{
Shows main information from \code{summary.fishery} objects.
}
\examples{
# Read a data base of fishery class
fisheryData = system.file("extdata", "fisheryData.csv", package = "imarpe")

# Produce a object of fishery class
landing  = getFishingData(file = fisheryData, type = "fisheryinfo", varType = "landing", sp = "caballa")
class(landing)

# Print the summary of the fishery class object
sumLanding = summary(landing)
print(sumLanding)

sumLanding = summary(landing, language = "english")
print(sumLanding, language = "english")

}

