% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-class.R
\name{report.cpue}
\alias{report.cpue}
\title{Report method for cpue objects}
\usage{
\method{report}{cpue}(x, format = "latex", tangle = FALSE, output = NULL,
  open = TRUE, daysToPlot = c(1, 8, 15, 22), textAxis2 = NULL,
  textAxis4 = NULL)
}
\arguments{
\item{x}{Object of \code{cpue} class.}

\item{daysToPlot}{If is a daily plot by default the x-axis show the first day of a week
(1, 8, 15, 22) but could be change to show a specific day or to show the all days of the
data with \code{all}. This is including in a vector form.}

\item{textAxis2}{The text of the x axis.}

\item{textAxis4}{The text of the y axis.}
}
\description{
Export a report of catch per unit effort (cpue) class.
}
\examples{
# Read a example of a data base
fisheryData = system.file("extdata", "fisheryData.csv", package = "imarpe")

# Produce a object of cpue class
cpue  = getFishingData(file = fisheryData, type = "cpue", varType = "cpue", sp = "caballa", efforType = "capacidad_bodega")
class(cpue)

#Produce the report showing all days on the x-axis (by default)
report(cpue)

#Produce the report showing only the day 15 on the x-axis
report(cpue, daysToPlot = "15")

}

