% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery-class.R
\name{report.fishery}
\alias{report.fishery}
\title{Report method for fishery objects}
\usage{
\method{report}{fishery}(x, type = "singled", format = "latex",
  tangle = FALSE, output = NULL, open = TRUE, daysToPlot = c(1, 8, 15,
  22), textAxis2 = NULL, textAxis4 = NULL)
}
\arguments{
\item{x}{Object of \code{fishery} class.}

\item{type}{Complete this.}

\item{daysToPlot}{If is a daily plot by default the x-axis show the first day of a week
(1, 8, 15, 22) but could be change to show a specific day or to show the all days of the
data with \code{all}. This is including in a vector form.}

\item{textAxis2}{The text of the x axis.}

\item{textAxis4}{The text of the y axis.}
}
\description{
Export a report of landing or fishing effort.
}
\examples{
# Read a example of a data base
fisheryData = system.file("extdata", "fisheryData.csv", package = "imarpe")

# Produce a object of fishery class
landing  = getFishingData(file = fisheryData, type = "fisheryinfo", varType = "landing", sp = "caballa")
class(landing)

#Produce the report showing all days on the x-axis (by default)
report(landing)

#Produce the report showing only the day 15 on the x-axis
report(landing, daysToPlot = "15")

}

