% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishingMonitoring-class.R
\name{report.fishingMonitoring}
\alias{report.fishingMonitoring}
\title{Report method for x objects}
\usage{
\method{report}{fishingMonitoring}(x, format = "latex", tangle = FALSE,
  output = NULL, open = TRUE)
}
\description{
x
}

