% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imarpe-main.R
\name{report}
\alias{report}
\title{Report method}
\usage{
report(x, format, output, ...)
}
\arguments{
\item{format}{The format to export the report.}

\item{output}{Folder where the report will be saved.}

\item{...}{Extra arguments passed to \code{\link{report}} function.}

\item{object}{Object of class \code{fishery}, \code{cpue} and \code{bitacora}.}
}
\value{
A report on specific format.
}
\description{
This function built a report for each class including on imarpe package.
}
\author{
Criscely Lujan-Paredes, \email{criscelylujan@gmail.com}.
}
