% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{seasonAnchovyNC}
\alias{seasonAnchovyNC}
\title{Peruvian port data.}
\format{A \code{data.frame} with six columns: the name of the port, the pattern to find the name
 of the port, the longitude and latitude of their names, the area (north, south and central) of
  ports and the importance of the port according the history of the fishery.}
\usage{
seasonAnchovyNC
}
\description{
Data with a list of ports around peruvian coast.
}
\references{
imarpe package vignette.
}

