% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{seasonAnchovyS}
\alias{seasonAnchovyS}
\title{Anchovy season for the south region.}
\format{A \code{data.frame} with four columns: year, the start date of the season for the
year, the end date of the season, and a code for the season.}
\usage{
seasonAnchovyS
}
\description{
Historical records of the opening and closing of anchovy fishing
seasons in the south region of the Peruvian coast.
}
\references{
imarpe package vignette.
}

