% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery-class.R
\name{summary.fishery}
\alias{summary.fishery}
\title{Summary method for fishery objects}
\usage{
\method{summary}{fishery}(object, language = "spanish")
}
\arguments{
\item{object}{Object of class \code{fishery}.}

\item{language}{The select language to print the summary of fishery objects.}
}
\value{
A \code{list} of summary.fishery class. This contains:
\itemize{
  \item var The type of the variable that has been to analyze. This can be lading or effort.
  \item portDay A data frame with the information of the variable analyzed by day and port.
  \item day A data frame with the information of the variable by day.
  \item port A data frame with the information of the variable by ports.
  \item months A data frame with the information of the variable by months.
  \item years A data frame with the information of the variable by years.
}
}
\description{
Get summary information of landing and fishing effort included on
 \code{fishery} objects.
}

